Refactored the preprocessing script names.
